+++ Module/Convert.RegistryRule/Methods.ps1
@@ -132,10 +132,10 @@ function Get-RegistryPathFromWindowsStig
                 Write-Verbose -Message "[$($MyInvocation.MyCommand.Name)] Fixed Leading Backslash : $path"
             }
 
-            # There is an edge case where the STIG has a typo and the path is writen with a space after \SOFTWARE\  (V-68819)
-            if  ($path -match '\sP' )
+            # There is an edge case where the STIG has a typo and the path is written with a space after \SOFTWARE\  (V-68819)
+            if  ($path -match '\\\sP.+\\')
             {
-                $path = $path -replace '\s'
+                $path = $path -replace '\\\sP', '\P'
             }
             $result += $path
         }
@@ -187,6 +187,41 @@ function Get-RegistryValueType
     }
 
     $return
+}
+
+<#
+    .SYNOPSIS
+        Tests that the ValueType is able to be used in a STIG
+
+    .PARAMETER TestValueType
+        The string to test against known good ValueTypes
+#>
+function Test-RegistryValueType
+{
+    [CmdletBinding()]
+    [OutputType([string])]
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [string]
+        $TestValueType
+    )
+
+    foreach ($valueType in $dscRegistryValueType.Keys)
+    {
+        if ($TestValueType -match $valueType)
+        {
+            $return = $valueType
+            break
+        }
+    }
+
+    if ($null -eq $return)
+    {
+        $return = $TestValueType
+    }
+
+    return $return
 }
 
 <#
@@ -865,10 +900,17 @@ function Test-MultipleRegistryEntries
     if (Test-SingleLineStigFormat -CheckContent $CheckContent)
     {
         $matches = $CheckContent | Select-String -Pattern "(HKLM|HKCU)\\" -AllMatches
+
+        if ($matches.Matches.Count -gt 1 -and $matches -match 'outlook\\security')
+        {
+            return $false
+        }
+
         if ( $matches.Matches.Count -gt 1 )
         {
             return $true
         }
+
         return $false
     }
     else
@@ -932,6 +974,7 @@ function Split-MultipleRegistryEntries
             {
                 $paths = $($CheckContent -join " ") -Split "AND(\s*)Procedure:"
             }
+
             if ( $CheckContent -match 'Navigate to:' )
             {
                 $keys = @()
@@ -942,6 +985,7 @@ function Split-MultipleRegistryEntries
                     {
                         $keys += $line
                     }
+
                     if ( $line -match 'REG_DWORD value' )
                     {
                         foreach ($key in $keys)
@@ -955,6 +999,10 @@ function Split-MultipleRegistryEntries
             }
         }
 
+        if ($paths.Count -lt 2)
+        {
+            $paths = $paths -split " and "
+        }
         foreach ($path in $paths)
         {
             if (![string]::IsNullOrWhiteSpace($path))